pyproject.toml
[project]
name = "sparv-ocr-suggestion-plugin"
version = "0.1.0"
description = "A sparv plugin for computing suggested OCR improvements."
authors = [
    { name = "Kristoffer Andersson", email = "kristoffer.andersson@gu.se" },
]
dependencies = ["sparv-pipeline >=5.2.0", "transformers>=4.34.1"]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9,<3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    # "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: Implementation :: CPython",
    # "uncomment if you test on these interpreters:",
    # "Programming Language :: Python :: Implementation :: IronPython",
    # "Programming Language :: Python :: Implementation :: Jython",
    # "Programming Language :: Python :: Implementation :: Stackless",
    "Topic :: Utilities",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points."sparv.plugin"]
ocr_suggestion = "ocr_suggestion"

[project.urls]
Homepage = "https://spraakbanken.gu.se"
Repository = "https://github.com/spraakbanken/sparv-ocr-suggestion-plugin"
"Bug Tracker" = "https://github.com/spraakbanken/sparv-ocr-suggestion-plugin/issues"

[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs"]

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_suggestion"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
target-version = ["py39", "py310", "py311"]

[tool.hatch.envs.test]
dependencies = ["pytest", "pytest-cov"]

# [tool.hatch.envs.test.scripts]
# run = "pytest tests"

[[tool.hatch.envs.test.matrix]]
python = ["3.9", "3.10", "3.11", "3.12"]

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.4",
    "mypy>=1.8.0",
    "bump2version>=1.0.1",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.14",
]
